requirements.txt
flask==3.0.0
gunicorn
requests==2.31.0
python-dotenv==1.0.0
langchain-community
faiss-cpu
fastembed
pypdf
groq

gspread==6.1.2
google-auth==2.35.0
